requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
decorator==5.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
loguru==0.6.0
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
shortuuid==1.0.11
sqlparse==0.4.3
urllib3==1.26.13
validators==0.20.0
